separate responsibility of callback from loader
